backend/tests/requirements.txt
requests==2.31.0
pytest==8.0.0
pytest-dotenv==0.5.2
pandas
openpyxl
pytest-mock
google-cloud-firestore
google-cloud-storage
google-cloud-texttospeech
